requirements.txt
Django==1.9.6
django-tastypie
python-mimeparse
pyyaml
requests
python-dateutil